+++ lib/Db/FeedMapperV2.php
@@ -52,13 +52,24 @@ class FeedMapperV2 extends NewsMapperV2
     public function findAllFromUser(string $userId, array $params = []): array
     {
         $builder = $this->db->getQueryBuilder();
-        $builder->select('feeds.*', $builder->func()->count('items.id', 'unreadCount'))
+        $builder
+            ->select(
+                'feeds.*',
+                $builder->func()->count('items_unread.id', 'unreadCount'),
+                $builder->func()->count('items_starred.id', 'starredCount')
+            )
             ->from(static::TABLE_NAME, 'feeds')
             ->leftJoin(
                 'feeds',
                 ItemMapperV2::TABLE_NAME,
-                'items',
-                'items.feed_id = feeds.id AND items.unread = :unread'
+                'items_unread',
+                'items_unread.feed_id = feeds.id AND items_unread.unread = :unread'
+            )
+            ->leftJoin(
+                'feeds',
+                ItemMapperV2::TABLE_NAME,
+                'items_starred',
+                'items_starred.feed_id = feeds.id AND items_starred.starred = 1'
             )
             ->where('feeds.user_id = :user_id')
             ->andWhere('feeds.deleted_at = 0')