+++ src/routes/index.ts
@@ -13,6 +13,7 @@ import store from './../store/app'
 export const ROUTES = {
 	EXPLORE: 'explore',
 	STARRED: 'starred',
+	STARREDFEED: 'starredfeed',
 	UNREAD: 'unread',
 	FEED: 'feed',
 	FOLDER: 'folder',
@@ -36,7 +37,16 @@ const getInitialRoute = function() {
 			params,
 		}
 	case '2':
-		return { name: ROUTES.STARRED }
+		if (store.state.lastViewedFeedId) {
+			params.feedId = store.state.lastViewedFeedId
+			return {
+				name: ROUTES.STARREDFEED,
+				params,
+			}
+		}
+		return {
+			name: ROUTES.STARRED,
+		}
 	case '3':
 		return { name: ROUTES.ALL }
 	case '5':
@@ -58,7 +68,18 @@ const routes = [
 		name: ROUTES.STARRED,
 		path: '/starred',
 		component: StarredPanel,
+		redirect: { name: 'starredItems' },
 		props: true,
+		children: [
+			{
+				name: 'starredItems',
+				path: '',
+			},
+			{
+				name: ROUTES.STARREDFEED,
+				path: ':feedId',
+			},
+		],
 	},
 	{
 		name: ROUTES.UNREAD,